make existing twitter_client.rb method-specific class
